Make the Player bounce off the walls and pop dots
I implemented most basic interactions between the Player node and other
objects: on contact with the map border (walls) the Player node will
bounce off them without losing velocity, and if the Player node
encounters pop-able dots, it will pop them. The Player node ceases to
exist after 3 bounces, which is a base for the future implementation of
the win condition.