Change the shooting mechanic to be bow-like
Now instead of left-clicking on a location to shoot in the direction of
the click, a player has to left-click on the Player node and drag and
hold the mouse cursor in the opposite direction from where they intend
to shoot, and, when they are ready, release the left mouse button to
perform a shot. Releasing the left click on top of the Player node does
not result in a shot, so it works as an aiming reset of sorts.